fx: favs

+++ src/modules/favorite/favorite.controller.ts
@@ -2,6 +2,8 @@ import { Controller, Delete, Get, HttpCode, Param, Post } from '@nestjs/common';
 import { FavoriteService } from './favorite.service';
 import { entities } from 'src/utils/entity';
 import { UUID } from 'src/utils/uuid';
+import { validate } from 'uuid';
+import { EntityNotContent, EntityNotCreate } from 'src/errors/errors';
 
 @Controller(entities.favs)
 export class FavoriteController {
@@ -18,7 +20,10 @@ export class FavoriteController {
 
   @HttpCode(204)
   @Delete('/:type/:id')
-  delete(@Param() { id }: UUID, @Param('type') type: string) {
-    return this.favService.delete(id, type);
+  async delete(@Param() { id }: UUID, @Param('type') type: string) {
+    const isValidUUID = validate(id);
+    if (!isValidUUID) throw new EntityNotCreate(type);
+    await this.favService.delete(id, type);
+    throw new EntityNotContent(type);
   }
 }
